Expose service version property
+ Expose the service version.
+ Remove code which attempted to forced all services to have the version
of the first service.  I believe the intent was to force all Zenoss
service versions to match the tenant service; however, it failed because
the root service is not necessarily the first service.  Also, it did not
allow the version to be updated (i.e. to reflect that the service had
been migrated)  Finalkly, it is incorrect to enforce that all
sub-services have the same version; it is a reasonable strategy to
version each service independently, and the service migration SDK should
not prevent this.