Do not migrate unmodified services
Previously, all services (whether modified or not) were included in the
list of "Modified" services for migration.  This put an unnecessary load
on the server in the case where the services tree was very large but the
list of changed services very small, with the result that service
migrations and zenpack install could timeout.  Following this change,
only services which were actually modified will be included.